move dds type from utils.h to dds/dds.h and update wrapper
