+++ src/lava/magma/runtime/message_infrastructure/message_infrastructure/csrc/channel/dds/dds.h
@@ -13,6 +13,30 @@
 #include <vector>
 
 namespace message_infrastructure {
+
+enum DDSTransportType {
+  DDSSHM = 0,
+  DDSTCPv4 = 1,
+  DDSTCPv6 = 2,
+  DDSUDPv4 = 3,
+  DDSUDPv6 = 4
+};
+
+enum DDSBackendType {
+  FASTDDSBackend = 0,
+  CycloneDDSBackend = 1
+};
+
+enum DDSInitErrorType {
+  DDSParticipantError = 1,
+  DDSPublisherError = 2,
+  DDSSubscriberError = 3,
+  DDSTopicError = 4,
+  DDSDataWriterError = 5,
+  DDSDataReaderError = 6,
+  DDSTypeParserError = 7
+};
+
 class DDSPublisher {
  public:
   virtual int Init() = 0;
