+++ src/lava/magma/runtime/message_infrastructure/message_infrastructure/csrc/channel_proxy.h
@@ -8,6 +8,9 @@
 #include <message_infrastructure/csrc/core/abstract_channel.h>
 #include <message_infrastructure/csrc/core/utils.h>
 #include <message_infrastructure/csrc/port_proxy.h>
+#if defined(DDS_CHANNEL)
+#include <message_infrastructure/csrc/channel/dds/dds.h>
+#endif
 
 #include <string>
